api-gateway/requirements.txt
fastapi
uvicorn
PyJWT
httpx
python-dotenv
prometheus_fastapi_instrumentator
